2ere pages OK, ajout complet de Account

+++ src/util/GestionContactUtils.java
@@ -4,6 +4,6 @@ public final class GestionContactUtils {
 	
 	private GestionContactUtils(){}
 	
-	public static final int BAD_ID = -1;
+	public static final long BAD_ID = -1;
 
 }
